pyproject.toml
[tool.poetry]
name = "main-project"
version = "0.1.0"
description = ""
authors = ["geonhui99 <leerjs99@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "==5.2.*"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.28.0"
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-dotenv = "^1.2.1"
drf-spectacular-sidecar = "^2025.10.1"
django-filter = "^25.2"
django-cors-headers = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
